scripts: genpulls.pl: treat first req as a coverletter
Add a command line option to treat the first request in a list as a
coverletter/summary.  It'll be identified as [GIT PULL 0/X]... where X
is the number of real requests.  eg 1 less than if --cover-letter isn't
specified.

As an example, I create a temporary branch 'merge-test', merge my
branches into it, sign and push.  I then list the merge-test branch as
the first request.  This provides an overall diffstat and commit summary
of what the series of requests contains.

It's purely informational, and there is no need for anyone to pull
merge-test.  Which is why it is 0/X.

Signed-off-by: Jason Cooper <jason@lakedaemon.net>