net: mv643xx_eth: update Device Tree bindings documentation
This patch updates the Device Tree bindings documentation for the
Marvell MV643xx ethernet controller node and makes the
"local-mac-address" and "phy" properties optional instead of required.
The absence of a "phy" node is legal and describes a MAC to switch
configuration, while not specifying the "local-mac-address" allows for
reuse of the MAC address already programmed into the Ethernet
controller.

From: Florian Fainelli <florian@openwrt.org>
Signed-off-by: Florian Fainelli <florian@openwrt.org>